+++ CHANGELOG.md
@@ -5,10 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [ 1.1.4 ] - 2024-06-29
+## [ 1.2.0 ] - 2024-06-29
 ### Added
 - Copy constructor to ProcessingProgress.
 - Added pure string constructor to ProcessingUnitProgressFormatter.
+- Added isEmpty in IProcessingUnitContext
+- Added toString methods to ProcessingUnitUtil.
+
+### Fixed
+- Nullpointer issues in case progress is null.
+- ProgressStatistic.toString had invalid characters.
 
 ## [ 1.1.2 ] - 2024-06-29
 ### Added